jely/requirements.txt
astroid==1.6.5
backports.functools-lru-cache==1.5
configparser==3.5.0
Django==1.11.13
django-cors-headers==2.2.0
djangorestframework==3.8.2
enum34==1.1.6
futures==3.2.0
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
mysqlclient==1.3.12
pylint==1.9.2
pytz==2018.4
singledispatch==3.4.0.3
six==1.11.0
wrapt==1.10.11
